refactor: 파싱 실패 뉴스레터를 모아두는 자료구조를 일급 컬렉션으로 수정
